refactor: (#8) Naming MysqlDatabase
