+++ src/main/kotlin/com/dsm/plugins/database/MysqlDataBaseFactory.kt
@@ -1,4 +1,4 @@
-package com.dsm.plugins
+package com.dsm.plugins.database
 
 import com.dsm.persistence.entity.AuthenticateStudentTable
 import com.dsm.persistence.entity.MissionTable
@@ -6,39 +6,36 @@ import com.dsm.persistence.entity.StudentTable
 import com.zaxxer.hikari.HikariConfig
 import com.zaxxer.hikari.HikariDataSource
 import io.ktor.server.config.ApplicationConfig
-import kotlinx.coroutines.Dispatchers
 import org.jetbrains.exposed.sql.Database
 import org.jetbrains.exposed.sql.SchemaUtils
 import org.jetbrains.exposed.sql.StdOutSqlLogger
 import org.jetbrains.exposed.sql.Table
-import org.jetbrains.exposed.sql.Transaction
 import org.jetbrains.exposed.sql.addLogger
-import org.jetbrains.exposed.sql.transactions.experimental.newSuspendedTransaction
 import org.jetbrains.exposed.sql.transactions.transaction
 
 /**
  *
- * 데이터베이스 연결을 담당하는 DataBaseFactory
+ * Mysql의 동작을 담당하는 MysqlDataBaseFactory
  *
  * @author Chokyunghyeon
  * @date 2023/03/16
  **/
-object DataBaseFactory {
+object MysqlDataBaseFactory {
 
-    private const val DB_PREFIX = "database"
-    private const val DB_URL = "$DB_PREFIX.url"
-    private const val DB_DRIVER = "$DB_PREFIX.driver"
-    private const val DB_USER = "$DB_PREFIX.user"
-    private const val DB_PASSWD = "$DB_PREFIX.password"
+    private const val MYSQL_PREFIX = "mysql"
+    private const val MYSQL_URL = "$MYSQL_PREFIX.url"
+    private const val MYSQL_DRIVER = "$MYSQL_PREFIX.driver"
+    private const val MYSQL_USER = "$MYSQL_PREFIX.user"
+    private const val MYSQL_PASSWD = "$MYSQL_PREFIX.password"
 
     operator fun invoke(config: ApplicationConfig) {
         val database = Database.connect(
             HikariDataSource(
                 HikariConfig().apply {
-                    driverClassName = config.property(DB_DRIVER).getString()
-                    jdbcUrl = config.property(DB_URL).getString()
-                    username = config.property(DB_USER).getString()
-                    password = config.property(DB_PASSWD).getString()
+                    driverClassName = config.property(MYSQL_DRIVER).getString()
+                    jdbcUrl = config.property(MYSQL_URL).getString()
+                    username = config.property(MYSQL_USER).getString()
+                    password = config.property(MYSQL_PASSWD).getString()
                 }
             )
         )
@@ -53,13 +50,5 @@ object DataBaseFactory {
             addLogger(StdOutSqlLogger)
             tables.run(SchemaUtils::create)
         }
-    }
-
-    suspend fun <T> dbQuery(database: Database? = null, block: suspend (Transaction) -> T): T = newSuspendedTransaction(
-        db = database,
-        context = Dispatchers.IO
-    ) {
-        addLogger(StdOutSqlLogger)
-        block(this)
     }
 }
